Switch to sampling arrays instead of lists to fix entropy calculation
Switch all the word lists, and character alternative lists to arrays
instead, and switch eRandomEltM to sample from arrays instead of
lists.

This should be more run-time efficient, especially for the sampling
the word list, which can be quite long, but more importantly its much
easier to properly account for the entropy introduced from a sample of
an array with random access versus a reservoir sampling algorithm on
lists.

Though the reservoir sampling algorithm, used to get a random element
from the list, properly selected from among the elements with equal
probability, it wasn't properly accounting for the entropy; each step
of the algorithm got a random number in the range [1,N] where N was
the number of elements seen so far, and each time this was introducing
log2(N) bits of entropy, except if it did one thing on 1, and
something else on every other result, meaning that results in [2,N]
actually yield much less entropy than a 1 result.

Indeed, choosing an element at random from a list of N items should
introduce entropy of log2(N), but the reservoir algorithm was
introducing sum([log2(n) | n <- 1..N]).

The length of the resulting words is noticeably longer.