Add more words and joining phrases to try to get shorter words
With more words, and more choices of joining strings, each word should
add more entropy. Extend the list of words to be between 3 and 10
characters long, and add a few more symbols to the join list.

Remove "!|1" from substitutions for "L", since they are already
substitutions for "I" and it makes the results even more confusing.